sql: move copy_in_test to pkg/sql/copy to have all copy tests together
Epic: CRDB-18892
Informs: #91831
Release note: None
