+++ pkg/sql/BUILD.bazel
@@ -580,7 +580,6 @@ go_test(
         "conn_executor_savepoints_test.go",
         "conn_executor_test.go",
         "conn_io_test.go",
-        "copy_in_test.go",
         "copy_test.go",
         "crdb_internal_test.go",
         "create_function_test.go",
@@ -826,8 +825,6 @@ go_test(
         "//pkg/util/shuffle",
         "//pkg/util/stop",
         "//pkg/util/syncutil",
-        "//pkg/util/timeofday",
-        "//pkg/util/timetz",
         "//pkg/util/timeutil",
         "//pkg/util/tracing",
         "//pkg/util/tracing/tracingpb",
