+++ pkg/sql/sqltestutils/BUILD.bazel
@@ -38,6 +38,7 @@ go_library(
         "//pkg/util/treeprinter",
         "@com_github_cockroachdb_datadriven//:datadriven",
         "@com_github_cockroachdb_errors//:errors",
+        "@com_github_jackc_pgx_v4//:pgx",
         "@com_github_lib_pq//:pq",
         "@com_github_stretchr_testify//require",
     ],
